Revert "Do not set content-type HTML by default"
This reverts commit 296989edd986f43069fd2648f747c34fc3e123d2.

There is a good reason to set HTML by default: compression will work on
resources without explicit content-type header. At the time the
compression wrapper is run the http package has not yet determined the
content-type heuristically so it doesnt know that the page is safe to
compress. Not pretty but okay I guess.