Switch secure cookie implementation
Previously, secure cookies in web.go were only cryptographically signed.
This prevented them from being tampered with. However, the contents of
the cookies were still transmitted in plain text to the client.

Instead of only signing the contents of the cookie, encrypt the contents
as well. This prevents any kind of information leakage.

Secure cookies are now encrypted with AES counter mode with a 32 bit
key. The contents are still signed using HMAC. Both the encryption key
and the signature key are generated using pbkdf2 using the CookieSecret
config option as the password source. The ciphertext, initialization
vector, and signature are now transmitted to the client.

Although the API is the same, cookies previously stored will not be
readable. Unfortunately there is no smooth upgrade process.

An example of using secure cookies has been added as well.

Fixes #160