Update the way default config is assigned
Previously, when a new web.Server object was created using the NewServer
method, the main config would be assigned as the server config. Instead,
extract the default config into a variable and use that. This is mainly
to allow server configs to be independent of the main server config if
necessary.